requirements.txt
requests==2.31.0
tenacity==8.2.3
pydantic==2.5.2
dbt-duckdb==1.7.1
